Update wordcloud to 1st draft and rearrange resume
The rearranged resume is hybrid style, and can be easily rearranged
in this commit.

Job and project descriptions updated for better key words and ATS
searchability.

Website link '~' sign was fixed for copying.

Dates were updated to have full name of months and year instead of
abbreviations. Hopefully that is good for ATS parsing.