Move wordcloud code into the skills section
The wordcloud is still shown at the top right corner of the resume,
but the text selection is now proper. From top, the starting text is
the name, address, etc., followed by summary, then skills, then the
wordcloud, followed by the remaining sections. This is in terms of
text selection.

Had to adjust top margin.

Had to add vspace to properly wrap text.

Changed the wordcloud word ordering roughly from top to bottom and
left to right.