Update wordcloud.tex to fix GitHub Action - 2
The previous commit did not fix the GitHub Action failure. The problem
may be due to the "convert" option supplied to \documentclass[]{}
instead. This is the actual source of -shell-escape warning mentioned
in the previous commit, which is needed for pdflatex to call
ImageMagick's "convert" to generate a *.png image of the wordcloud.
Since we don't need the png file in the GitHub Action build, I'm
commenting out the convert option. Let's see if this fixes the GitHub
Action failure.